fix: 하드웨어까지 test 시 푸시알림 뜨지 않는 문제 해결 (#56)
