+++ public/firebase-messaging-sw.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-undef */
 // 이 파일은 Service Worker로 동작하므로 번들러(Vite)를 거치지 않는 정적 파일이다.
 // 따라서 import.meta.env / ES module import를 쓸 수 없고, compat SDK를 importScripts로 불러온다.
 // SDK 버전은 package.json의 firebase 버전(12.14.0)과 반드시 일치시킨다.
@@ -19,10 +18,11 @@ firebase.initializeApp({
 const messaging = firebase.messaging();
 
 // 앱이 백그라운드(탭 비활성/닫힘) 상태일 때 도착하는 푸시 메시지를 처리한다.
-// 백엔드는 data-only 페이로드를 보낸다(notification 키 없음). data-only는 브라우저가
-// 알림을 자동 표시하지 않으므로 여기서 직접 showNotification으로 띄운다.
-// 이렇게 해야 중복 표시(자동 1개 + 수동 1개)가 생기지 않고, notificationclick 핸들러도 항상 동작한다.
+// notification 페이로드는 브라우저가 자동 표시하므로 수동으로 다시 띄우지 않는다.
+// data-only 페이로드만 직접 표시해 두 형식을 모두 지원하면서 중복 알림을 방지한다.
 messaging.onBackgroundMessage((payload) => {
+  if (payload.notification) return;
+
   const { title, body } = payload.data ?? {};
   self.registration.showNotification(title ?? '알림', {
     body,
